[cheng.gao] #calltrace: adjust folder structure

(cherry picked from commit a899bf5c2c670dd6c5e25f615878251333fc209b)
