+++ SourceCode/Arch/arm/include/arm/call_trace.h
@@ -1,0 +1,10 @@
+//
+// Created by XingfengYang on 2021/1/8.
+//
+
+#ifndef SYNESTIAOS_CALL_TRACE_H
+#define SYNESTIAOS_CALL_TRACE_H
+
+void dump_calltrace(void);
+
+#endif //SYNESTIAOS_CALL_TRACE_H
