+++ SourceCode/Kernel/include/kernel/assert.h
@@ -5,6 +5,9 @@
 #ifndef SYNESTIAOS_ASSERT_H
 #define SYNESTIAOS_ASSERT_H
 
+#include "arm/interrupt.h"
+#include "arm/call_trace.h"
+
 void __assert_fail(const char *file, int line, const char *failedExpr);
 
 void __assert_fail_msg(const char *file, int line, const char *failedExpr, const char *msg);
@@ -13,6 +16,8 @@ void __assert_fail_msg(const char *file, int line, const char *failedExpr, const
     do {                                           \
         if (!(x)) {                                \
             __assert_fail(__FILE__, __LINE__, #x); \
+            dump_calltrace();                      \
+            dead();                                \
         }                                          \
     } while (0)
 
@@ -20,6 +25,8 @@ void __assert_fail_msg(const char *file, int line, const char *failedExpr, const
     do {                                                    \
         if (!(x)) {                                         \
             __assert_fail_msg(__FILE__, __LINE__, #x, msg); \
+            dump_calltrace();                               \
+            dead();                                         \
         }                                                   \
     } while (0)
